Refactor backend filenames of qc_v1_parser.py to parser.py and qc_v1_compiler.py to compiler.py
